requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.6
djangorestframework==3.10.3
et-xmlfile==1.0.1
geographiclib==1.50
geopy==1.20.0
httpie==1.0.3
idna==2.8
jdcal==1.4.1
lxml==4.4.1
openpyxl==3.0.0
Pygments==2.4.2
PyPDF2==1.26.0
pyperclip==1.7.0
pytz==2019.3
requests==2.22.0
soupsieve==1.9.4
sqlparse==0.3.0
termcolor==1.1.0
urllib3==1.25.6
xlwt==1.3.0
